fix(client) : FilteringButton에서 익명함수 사용 제거
